feat: make bg option border remain after selection
Before, the background color options only shoed a border on mouseenter
and reset it on mouseleave. Updated the logic so that clicking an option
highlights it by setting its border color, and removes the border
highlight from other options. This gives the user clear feedback about
which background is currently selected.